+++ src/components/MyHome/UserPoster/index.jsx
@@ -4,20 +4,63 @@ import MyhomeModal from '../../Modal/MyhomeModal';
 import UserPosterItem from '../UserPosteItem';
 import { ModalTitle, ModalTitleButton, ModalTitleWrapper } from './style';
 
-import { useState } from 'react';
-import { usePoster } from '../../../hooks/usePoster';
+import { useEffect, useState } from 'react';
+import { useRecoilState } from 'recoil';
+import { getUser } from '../../../apis/auth';
+import {
+  myHomeModalState,
+  uploadImageState,
+  userLoginDateState,
+} from '../../../recoil/uploadImageState';
 import Pagination from '../../Pagination';
 
 const LoginPoster = () => {
-  const {
-    handlePoster,
-    handleDeletePoster,
-    handlerModalClose,
-    postId,
-    visible,
-    imageValue,
-    posts,
-  } = usePoster();
+  const [postId, setPostId] = useState('');
+  const [visible, setVisible] = useRecoilState(myHomeModalState);
+  const [postDetail, setPostDetail] = useRecoilState(userLoginDateState);
+  const [imageValue, setImageValue] = useRecoilState(uploadImageState);
+  const [posts, setPosts] = useState([]);
+  const [loading, setLoading] = useState(false);
+
+  const handlePoster = async (id) => {
+    const getPostDetail = await getMyPostDetail(id);
+    setVisible(true);
+    setPostId(id);
+    setPostDetail(getPostDetail);
+  };
+
+  const handlerModalClose = () => {
+    setVisible(false);
+    setImageValue(null);
+  };
+
+  const handleDeletePoster = async (id) => {
+    swal({
+      title: USER.DELETE_POSTER,
+      text: USER.DELETE,
+      icon: ERROR_MESSAGE_SIGNIN.DELETE_POSTER_WARNING,
+      dangerMode: true,
+    }).then(async (willDelete) => {
+      if (willDelete) {
+        await removePost(id);
+
+        setPosts({
+          ...posts,
+          posts: posts.filter((post) => post._id !== id),
+        });
+      }
+    });
+  };
+
+  useEffect(() => {
+    const getUserData = async () => {
+      const getLoginUserData = await getUser();
+
+      setPosts(getLoginUserData.data.posts);
+      setLoading(false);
+    };
+    getUserData();
+  }, []);
 
   const [currentPage, setCurrentPage] = useState(1);
   const [postPerPage, setPostPerPage] = useState(3);
